requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.12.2
boto3==1.28.40
botocore==1.31.40
certifi==2023.7.22
charset-normalizer==3.2.0
discord==2.3.2
discord.py==2.5.2
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
jmespath==1.0.1
multidict==6.0.4
numpy==1.25.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.1.0
psycopg2==2.9.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
s3transfer==0.6.2
selenium==4.31.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.20
tqdm==4.66.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2023.3
urllib3==1.26.16
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.2
